added py spike unit test; deprecated unused simulation py test file
